topic9/spring-security-db/src/main/java/com/kma/practice8/springsecuritydb/controller/WebController.java
package com.kma.practice8.springsecuritydb.controller;

import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class WebController {

    @GetMapping("/")
    public String index() {
        return "index";
    }

    @GetMapping("/admin")
    public ResponseEntity<String> admin() {
        return ResponseEntity.ok("admin_root");
    }

    @GetMapping("/admin/subpage")
    public String adminSubpage() {
        return "admin_sub";
    }

    @GetMapping("/catalog")
    public String catalog() {
        return "catalog";
    }

    @GetMapping("/profile")
    public String profile() {
        return "profile";
    }

    @GetMapping("/other")
    public String other() {
        return "other";
    }

}
